Transfer plot into separate file
